requirements.txt
tkinterdnd2
moviepy
openai-whisper
faster-whisper
SpeechRecognition
imageio-ffmpeg
pillow
numpy
scipy
psutil
requests
customtkinter
opencv-python
